Task 2. Explore the useradd command
In this task, imagine that you want to set the expiration date for a temporary user account. You know that you need to use the useradd command for this, but you’re not quite sure how to complete the task. You realize it might involve adding an option to the command.

1.Use the most appropriate Linux command to get help on the useradd command and learn more about all of its options.
2.Press Q to exit this manual page.