Task 4. Determine which command to use
In this task, imagine that you need to create a new group but you can’t remember what command to use. You need to identify a command that will do this by searching for it through keywords. In this case, use the keywords create new group.

Use the most appropriate Linux command with these keywords to identify what command to use.